Thermalization fixed?
I think this might finally be correct. We have to adjust the signs on the off-diagonal terms of T1 depending on the signs of the eigenvectors (we just check the signs of Lx^2+Ly^2+Lz^2 to figure out the relative signs). However, it appears we also need to adjust the signs of the thermalization step. I think this is maybe right finally.

Also allow 'e' to be used to specify e- for delta-pulse