Hz axis definition updates (plus version info)
Adjusts the behavior of the Hz axis when plotting nuclei with negative gyromagnetic ratios (ppm and Hz axes now run in the same direction, to be consistent with SIMPSON). This also adjusts behavior when defining CS and CSA Hamiltonians using Hz. Also added version_info, which is text describing updates to the current version.